Add code to fetch a list of vehicles
Fetches vehicle make, model and year from a javascript file on
cars.com. It stores the information into the database. On subsequent
fetches, it should update the database so that it knows that the
vehicle is still active. This can be run manually by administrators,
and is also run as a cron job every Monday at midnight.