Store the server response when grabbing car information
We will now store the response from cars.com so that we can see if the
result has changed since the last time we pulled for data. If it has
changed, we should be able to compare the JSONArrays and see what
exactly changed without having to run a Database query on each vehicle,
even if the Vehicle hasn't changed. In doing this, the cron job has
been changed to run every day at midnight, because most of the time
there won't be any changes, so it won't hurt.