Add some methods to models to get formatted data
For consistency, I have added methods to some of the models to get
their data as formatted strings so that when we are displaying them on
the site, they are consistent between pages (for instance, odometer
readings should be ints without any floating point part, and should
have thousands separators).  Still a work in progress.