Change how photos work
Each record that contains an image will also have a pictureurl property
that contains the url for that image, scaled properly. This removes the
need for the blob.py module.

Also, (somewhat) restrict file uploads to a small number of image
types.  This is based on the browser's implementation of the file input
type, but it is better than nothing.

Issues #5 and #2